check checker

+++ philo_one/srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: mroux <mroux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/08 21:55:44 by mroux             #+#    #+#             */
-/*   Updated: 2021/03/25 20:04:47 by mroux            ###   ########.fr       */
+/*   Updated: 2021/03/25 20:08:26 by mroux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ void			start_checker(t_global *gl)
 {
 	int				i;
 	struct timeval	now;
-	unsigned long 	time;
+	long 	time;
 	int				status_meal;
 
 	while (1)
@@ -90,7 +90,7 @@ void			start_checker(t_global *gl)
 		while (i < gl->number_of_philos)
 		{
 			gettimeofday(&now, NULL);
-			if ((time = (timeval_to_ms(&now) - timeval_to_ms(&gl->philos[i].last_lunch))) >= gl->time_to_die)
+			if ((time = (long) (timeval_to_ms(&now) - timeval_to_ms(&gl->philos[i].last_lunch))) >= (long) gl->time_to_die)
 			{
 				printf("==%lu==\n", timeval_to_ms(&now) - timeval_to_ms(&gl->philos[i].last_lunch));
 				printf("==%ld==\n", (long)(timeval_to_ms(&now) - timeval_to_ms(&gl->philos[i].last_lunch)));
